Photon changes part 2: Function registry and type system changes (#3)
This patch changes the function registry to support any accepted input type for an expression. Before, each function was registered with a particular set if input and output types: for example, the `If` expression could only be used with the output type `int` and the input type `long`. This was very constraining, since expressions could only be used in a very limited contexts.

Now, the function registry contains one entry for every valid type combination. For example, the Pow expression can accept any numeric value, instead of a particular numeric value chosen at startup.

This patch also introduces the concept of "generics," which allows us to set a return type, and derive input types from it. This adds additional flexibility, and allows us to express expressions that can take any type (with constraints on parameters that need to be the same type) without exhaustively enumerating them. As an example, the CASE WHEN expression has the signature (Boolean, A, Boolean, A, ...), where A can be any type (as long as all A's in the expression are the same type). We need generics for _nested types_: if were just handling atomic types, it would have been simpler to just enumerate all combinations explicitly, e.g.,

```scala
// Coalesce that takes one input and returns one output. Without nested types, we could have just enumerated all types.
// However, with nested types, there's not way to register all types in the registry. Generics solve that.
registerFunction("coalesce", Int, (Int)),
registerFunction("coalesce", Short, (Short)),

// With generics. Any input and output type, as long as they match
registerFunction("coalesce", Generics.A, (Generics.A)),
```

Other changes in this patch:

* Changes random type generation to avoid code duplication for nested types. This also makes the nested types generation a bit more consistent by following the RQG configs.

* Changes data types to not generate random fields when they are constructed. This makes their use more predictable. Use `RandomUtils` to generate random types instead.

* Fixes a bug in the expression registry where we were using the runtime class constructor list to populate the registry. The order of constructors returned by `runtimeClass.getConstructors` was non-deterministic, so this would lead to non-determinism for the same seed. We now always choose the constructor with the maximum parameter count.